Consolidate project files into core classes and remove unused components
