--- Core/RobloxOffsets.cs
@@ -1,166 +1,0 @@
-using System;
-
-namespace SelectXYZ_Cheat.Core
-{
-    // Roblox Version: version-b8550645b8834e8a
-    public static class RobloxOffsets
-    {
-        public const uintptr_t Adornee = 0xD8;
-        public const uintptr_t AnchoredMask = 0x2;
-        public const uintptr_t AnimationId = 0xD8;
-        public const uintptr_t AttributeList = 0x138;
-        public const uintptr_t AttributeToNext = 0x70;
-        public const uintptr_t AttributeToValue = 0x30;
-        public const uintptr_t AutoJumpEnabled = 0x1E3;
-        public const uintptr_t BeamBrightness = 0x198;
-        public const uintptr_t BeamColor = 0x128;
-        public const uintptr_t BeamLightEmission = 0x1A4;
-        public const uintptr_t BeamLightInfuence = 0x1A8;
-        public const uintptr_t Camera = 0x428;
-        public const uintptr_t CameraMaxZoomDistance = 0x2B8;
-        public const uintptr_t CameraMinZoomDistance = 0x2BC;
-        public const uintptr_t CameraMode = 0x2C0;
-        public const uintptr_t CameraPos = 0x124;
-        public const uintptr_t CameraRotation = 0x100;
-        public const uintptr_t CameraSubject = 0xF0;
-        public const uintptr_t CameraType = 0x160;
-        public const uintptr_t CanCollideMask = 0x8;
-        public const uintptr_t CanTouchMask = 0x10;
-        public const uintptr_t CharacterAppearanceId = 0x260;
-        public const uintptr_t Children = 0x80;
-        public const uintptr_t ChildrenEnd = 0x8;
-        public const uintptr_t ClassDescriptor = 0x18;
-        public const uintptr_t ClassDescriptorToClassName = 0x8;
-        public const uintptr_t ClickDetectorMaxActivationDistance = 0x118;
-        public const uintptr_t ClockTime = 0x1C0;
-        public const uintptr_t CreatorId = 0x190;
-        public const uintptr_t DataModelDeleterPointer = 0x6ED6E40;
-        public const uintptr_t DataModelPrimitiveCount = 0x410;
-        public const uintptr_t DataModelToRenderView1 = 0x1D8;
-        public const uintptr_t DataModelToRenderView2 = 0x8;
-        public const uintptr_t DataModelToRenderView3 = 0x28;
-        public const uintptr_t DecalTexture = 0x1A0;
-        public const uintptr_t Deleter = 0x10;
-        public const uintptr_t DeleterBack = 0x18;
-        public const uintptr_t Dimensions = 0x720;
-        public const uintptr_t DisplayName = 0x118;
-        public const uintptr_t EvaluateStateMachine = 0x1E3;
-        public const uintptr_t FOV = 0x168;
-        public const uintptr_t FakeDataModelPointer = 0x6ED6E38;
-        public const uintptr_t FakeDataModelToDataModel = 0x1C0;
-        public const uintptr_t FogColor = 0x104;
-        public const uintptr_t FogEnd = 0x13C;
-        public const uintptr_t FogStart = 0x140;
-        public const uintptr_t ForceNewAFKDuration = 0x1F8;
-        public const uintptr_t FramePositionOffsetX = 0x3C4;
-        public const uintptr_t FramePositionOffsetY = 0x3CC;
-        public const uintptr_t FramePositionX = 0x3C0;
-        public const uintptr_t FramePositionY = 0x3C8;
-        public const uintptr_t FrameRotation = 0x190;
-        public const uintptr_t FrameSizeX = 0x120;
-        public const uintptr_t FrameSizeY = 0x124;
-        public const uintptr_t GameId = 0x198;
-        public const uintptr_t GameLoaded = 0x680;
-        public const uintptr_t Gravity = 0x968;
-        public const uintptr_t Health = 0x19C;
-        public const uintptr_t HealthDisplayDistance = 0x2E0;
-        public const uintptr_t HipHeight = 0x1A8;
-        public const uintptr_t HumanoidDisplayName = 0xD8;
-        public const uintptr_t HumanoidState = 0x870;
-        public const uintptr_t HumanoidStateId = 0x20;
-        public const uintptr_t InputObject = 0x108;
-        public const uintptr_t InsetMaxX = 0x108;
-        public const uintptr_t InsetMaxY = 0x10C;
-        public const uintptr_t InsetMinX = 0x100;
-        public const uintptr_t InsetMinY = 0x104;
-        public const uintptr_t InstanceCapabilities = 0x380;
-        public const uintptr_t JobEnd = 0x1D8;
-        public const uintptr_t JobId = 0x140;
-        public const uintptr_t JobStart = 0x1D0;
-        public const uintptr_t Job_Name = 0x18;
-        public const uintptr_t JobsPointer = 0x6fa7280;
-        public const uintptr_t JumpPower = 0x1B8;
-        public const uintptr_t LocalPlayer = 0x128;
-        public const uintptr_t LocalScriptByteCode = 0x1B0;
-        public const uintptr_t LocalScriptBytecodePointer = 0x10;
-        public const uintptr_t LocalScriptHash = 0x1C0;
-        public const uintptr_t MaterialType = 0x0;
-        public const uintptr_t MaxHealth = 0x1BC;
-        public const uintptr_t MaxSlopeAngle = 0x1C0;
-        public const uintptr_t MeshPartColor3 = 0xADA8;
-        public const uintptr_t ModelInstance = 0x328;
-        public const uintptr_t ModuleScriptByteCode = 0x158;
-        public const uintptr_t ModuleScriptBytecodePointer = 0x10;
-        public const uintptr_t ModuleScriptHash = 0x180;
-        public const uintptr_t MoonTextureId = 0xE0;
-        public const uintptr_t MousePosition = 0xF4;
-        public const uintptr_t MouseSensitivity = 0x6f4d9b4;
-        public const uintptr_t MoveDirection = 0x160;
-        public const uintptr_t Name = 0x78;
-        public const uintptr_t NameDisplayDistance = 0x2EC;
-        public const uintptr_t NameSize = 0x10;
-        public const uintptr_t OnDemandInstance = 0x30;
-        public const uintptr_t OutdoorAmbient = 0x110;
-        public const uintptr_t Parent = 0x50;
-        public const uintptr_t PartSize = 0x25C;
-        public const uintptr_t Ping = 0xD0;
-        public const uintptr_t PlaceId = 0x1A0;
-        public const uintptr_t PlayerConfigurerPointer = 0x6EB4238;
-        public const uintptr_t PlayerMouse = 0xC00;
-        public const uintptr_t Position = 0x14C;
-        public const uintptr_t Primitive = 0x178;
-        public const uintptr_t PrimitiveGravity = 0x120;
-        public const uintptr_t PrimitiveValidateValue = 0x6;
-        public const uintptr_t PrimitivesPointer1 = 0x3B0;
-        public const uintptr_t PrimitivesPointer2 = 0x210;
-        public const uintptr_t ProximityPromptActionText = 0xD8;
-        public const uintptr_t ProximityPromptEnabled = 0x14A;
-        public const uintptr_t ProximityPromptGamepadKeyCode = 0x134;
-        public const uintptr_t ProximityPromptHoldDuraction = 0x138;
-        public const uintptr_t ProximityPromptMaxActivationDistance = 0x140;
-        public const uintptr_t ProximityPromptMaxObjectText = 0xF8;
-        public const uintptr_t RenderJobToDataModel = 0x1B0;
-        public const uintptr_t RenderJobToFakeDataModel = 0x38;
-        public const uintptr_t RenderJobToRenderView = 0x218;
-        public const uintptr_t RequireBypass = 0x7E8;
-        public const uintptr_t RigType = 0x1D0;
-        public const uintptr_t Rotation = 0x130;
-        public const uintptr_t RunContext = 0x150;
-        public const uintptr_t ScriptContext = 0x3C0;
-        public const uintptr_t Sit = 0x1E3;
-        public const uintptr_t SkyboxBk = 0x108;
-        public const uintptr_t SkyboxDn = 0x130;
-        public const uintptr_t SkyboxFt = 0x158;
-        public const uintptr_t SkyboxLf = 0x180;
-        public const uintptr_t SkyboxRt = 0x1A8;
-        public const uintptr_t SkyboxUp = 0x1D0;
-        public const uintptr_t SoundId = 0xE0;
-        public const uintptr_t StarCount = 0x228;
-        public const uintptr_t StringLength = 0x10;
-        public const uintptr_t SunTextureId = 0x1F8;
-        public const uintptr_t TagList = 0x120;
-        public const uintptr_t TaskSchedulerMaxFPS = 0x1B0;
-        public const uintptr_t TaskSchedulerPointer = 0x6FA70A0;
-        public const uintptr_t Team = 0x248;
-        public const uintptr_t TeamColor = 0xD8;
-        public const uintptr_t Tool_Grip_Position = 0x45C;
-        public const uintptr_t Transparency = 0xF8;
-        public const uintptr_t UserId = 0x270;
-        public const uintptr_t Value = 0xD8;
-        public const uintptr_t Velocity = 0x158;
-        public const uintptr_t ViewportSize = 0x300;
-        public const uintptr_t VisualEngine = 0x10;
-        public const uintptr_t VisualEnginePointer = 0x6D18608;
-        public const uintptr_t VisualEngineToDataModel1 = 0x700;
-        public const uintptr_t VisualEngineToDataModel2 = 0x1C0;
-        public const uintptr_t WalkSpeed = 0x1DC;
-        public const uintptr_t WalkSpeedCheck = 0x3B8;
-        public const uintptr_t WhitelistedThreads = 0x0; // Encrypted
-        public const uintptr_t Workspace = 0x180;
-        public const uintptr_t WorkspaceToWorld = 0x3B0;
-        public const uintptr_t ViewMatrix = 0x4B0;
-    }
-
-    // C# doesn't have uintptr_t, so we use UIntPtr
-    using uintptr_t = UIntPtr;
-}
