+++ ESP.cs
@@ -1,12 +1,197 @@
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Numerics;
 using System.Threading.Tasks;
-using SelectXYZ_Cheat.Core;
-using SelectXYZ_Cheat.Models;
 
-namespace SelectXYZ_Cheat.ESP
+namespace SelectXYZ_Cheat
 {
+    #region ESP Settings
+    public enum ESPBoxType
+    {
+        None,
+        Box2D,
+        Box3D,
+        CornerBox
+    }
+
+    public class ESPSettings : INotifyPropertyChanged
+    {
+        private ESPBoxType _boxType = ESPBoxType.None;
+        private bool _skeletonESP = false;
+        private bool _healthBar = false;
+        private bool _healthNumber = false;
+        private bool _playerName = false;
+        private bool _distanceESP = false;
+        private bool _headCircle = false;
+        private bool _sonarESP = false;
+        private bool _chinaHat = false;
+        private bool _chams = false;
+        private bool _boxGlow = false;
+        private bool _boxFill = false;
+
+        // Colors
+        private string _skeletonColor = "White";
+        private string _cornerBoxColor = "White";
+        private string _healthBarColor = "Green";
+        private string _boxFillColor = "Black";
+        private string _headCircleColor = "White";
+        private string _box3DColor = "Black";
+        private string _outlineColor = "#FF8080FF";
+        private string _chamsColor = "#FF00FFFF";
+        private string _chinaHatColor = "Red";
+        private string _sonarColor = "#FF00FFFF";
+        private string _boxColor = "White";
+
+        public event PropertyChangedEventHandler? PropertyChanged;
+
+        public ESPBoxType BoxType
+        {
+            get => _boxType;
+            set { _boxType = value; OnPropertyChanged(nameof(BoxType)); }
+        }
+
+        public bool SkeletonESP
+        {
+            get => _skeletonESP;
+            set { _skeletonESP = value; OnPropertyChanged(nameof(SkeletonESP)); }
+        }
+
+        public bool HealthBar
+        {
+            get => _healthBar;
+            set { _healthBar = value; OnPropertyChanged(nameof(HealthBar)); }
+        }
+
+        public bool HealthNumber
+        {
+            get => _healthNumber;
+            set { _healthNumber = value; OnPropertyChanged(nameof(HealthNumber)); }
+        }
+
+        public bool PlayerName
+        {
+            get => _playerName;
+            set { _playerName = value; OnPropertyChanged(nameof(PlayerName)); }
+        }
+
+        public bool DistanceESP
+        {
+            get => _distanceESP;
+            set { _distanceESP = value; OnPropertyChanged(nameof(DistanceESP)); }
+        }
+
+        public bool HeadCircle
+        {
+            get => _headCircle;
+            set { _headCircle = value; OnPropertyChanged(nameof(HeadCircle)); }
+        }
+
+        public bool SonarESP
+        {
+            get => _sonarESP;
+            set { _sonarESP = value; OnPropertyChanged(nameof(SonarESP)); }
+        }
+
+        public bool ChinaHat
+        {
+            get => _chinaHat;
+            set { _chinaHat = value; OnPropertyChanged(nameof(ChinaHat)); }
+        }
+
+        public bool Chams
+        {
+            get => _chams;
+            set { _chams = value; OnPropertyChanged(nameof(Chams)); }
+        }
+
+        public bool BoxGlow
+        {
+            get => _boxGlow;
+            set { _boxGlow = value; OnPropertyChanged(nameof(BoxGlow)); }
+        }
+
+        public bool BoxFill
+        {
+            get => _boxFill;
+            set { _boxFill = value; OnPropertyChanged(nameof(BoxFill)); }
+        }
+
+        // Color Properties
+        public string SkeletonColor
+        {
+            get => _skeletonColor;
+            set { _skeletonColor = value; OnPropertyChanged(nameof(SkeletonColor)); }
+        }
+
+        public string CornerBoxColor
+        {
+            get => _cornerBoxColor;
+            set { _cornerBoxColor = value; OnPropertyChanged(nameof(CornerBoxColor)); }
+        }
+
+        public string HealthBarColor
+        {
+            get => _healthBarColor;
+            set { _healthBarColor = value; OnPropertyChanged(nameof(HealthBarColor)); }
+        }
+
+        public string BoxFillColor
+        {
+            get => _boxFillColor;
+            set { _boxFillColor = value; OnPropertyChanged(nameof(BoxFillColor)); }
+        }
+
+        public string HeadCircleColor
+        {
+            get => _headCircleColor;
+            set { _headCircleColor = value; OnPropertyChanged(nameof(HeadCircleColor)); }
+        }
+
+        public string Box3DColor
+        {
+            get => _box3DColor;
+            set { _box3DColor = value; OnPropertyChanged(nameof(Box3DColor)); }
+        }
+
+        public string OutlineColor
+        {
+            get => _outlineColor;
+            set { _outlineColor = value; OnPropertyChanged(nameof(OutlineColor)); }
+        }
+
+        public string ChamsColor
+        {
+            get => _chamsColor;
+            set { _chamsColor = value; OnPropertyChanged(nameof(ChamsColor)); }
+        }
+
+        public string ChinaHatColor
+        {
+            get => _chinaHatColor;
+            set { _chinaHatColor = value; OnPropertyChanged(nameof(ChinaHatColor)); }
+        }
+
+        public string SonarColor
+        {
+            get => _sonarColor;
+            set { _sonarColor = value; OnPropertyChanged(nameof(SonarColor)); }
+        }
+
+        public string BoxColor
+        {
+            get => _boxColor;
+            set { _boxColor = value; OnPropertyChanged(nameof(BoxColor)); }
+        }
+
+        protected virtual void OnPropertyChanged(string propertyName)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+    }
+    #endregion
+
+    #region ESP Manager
     public class ESPManager
     {
         private static ESPManager? _instance;
@@ -16,10 +201,10 @@ namespace SelectXYZ_Cheat.ESP
         private readonly List<Player> players = new();
         private readonly object playersLock = new();
 
-        public ESPSettings Settings { get; } = new ESPSettings();
-
         private MemoryReader? externalMemoryReader;
         private RobloxMemoryScanner? robloxScanner;
+
+        public ESPSettings Settings { get; } = new ESPSettings();
 
         private ESPManager()
         {
@@ -104,7 +289,6 @@ namespace SelectXYZ_Cheat.ESP
         private List<Player> SimulatePlayerDetection()
         {
             // This simulates finding players for demonstration
-            // Replace with actual memory reading logic
             var simulatedPlayers = new List<Player>();
 
             for (int i = 0; i < 3; i++)
@@ -136,9 +320,6 @@ namespace SelectXYZ_Cheat.ESP
         private Vector2 WorldToScreen(Vector3 worldPos)
         {
             // Simplified world-to-screen conversion
-            // In reality, you'd need the view matrix from the game
-            
-            // For simulation, just convert 3D position to 2D screen coordinates
             float screenX = 400 + (worldPos.X - 100) * 2;
             float screenY = 300 - (worldPos.Y - 50) * 3;
             
@@ -148,8 +329,8 @@ namespace SelectXYZ_Cheat.ESP
         private BoundingBox2D CalculateBoundingBox(Player player)
         {
             // Calculate 2D bounding box based on distance and screen position
-            float distance = player.Distance;
-            float scale = Math.Max(0.5f, 100.0f / distance); // Scale based on distance
+            float distance = Math.Max(player.Distance, 1.0f);
+            float scale = Math.Max(0.5f, 100.0f / distance);
             
             float width = 40 * scale;
             float height = 80 * scale;
@@ -165,12 +346,10 @@ namespace SelectXYZ_Cheat.ESP
 
         public bool IsPlayerInFOV(Player player, float fovAngle = 90f)
         {
-            // Simple FOV check - in reality you'd use camera direction and position
-            var screenCenter = new Vector2(960, 540); // Assuming 1920x1080
+            var screenCenter = new Vector2(960, 540);
             var distance = Vector2.Distance(player.ScreenPosition, screenCenter);
             
-            // Simple distance-based FOV check
-            return distance < 500; // Pixels from center
+            return distance < 500;
         }
 
         public Player? GetClosestPlayer()
@@ -223,4 +402,5 @@ namespace SelectXYZ_Cheat.ESP
             }
         }
     }
+    #endregion
 }
